API: Move crl to gencrl (it was already named after gencrl anyway)
